# GODRIVER-2025 Stop replacing the bsonrw.valueWriter buffer when given a SliceWriter destination.